Fix sign flag using the unsigned value for checking sign, add 0x00 ADD instruction, add 0x3C CMP instruction,add 0x52 PUSH instruction, add 0xAC LODSB instruction, add 0xBB MOV instruction, add 0xEA LJMP instruction, fix error reporting on 0x01 ADD for 66 prefix use, use new segment offsetting syntax for 0x8B MOV, 0x8E MOV, rename 0x75 JNE to JNE or JNZ, add hack to 0xC3 RET, disc reading/teletype work on 0xCD INT, simplify 0xE8 CALL, simplify 0xEB JMP, add new SegmentRegister class, move boot sector/primary loading to ISO9660Disc.kt

+++ src/main/kotlin/org/bread_experts_group/breadmod/experimental/computer/ia32/instruction/DecodingUtil.kt
@@ -118,7 +118,7 @@ class DecodingUtil(private val processor: IA32Processor) {
 	}
 
 	fun getFlagForResult(flag: IA32Processor.FlagType, value: ULong): Boolean = when (flag) {
-		IA32Processor.FlagType.SIGN_FLAG   -> value < 0u
+		IA32Processor.FlagType.SIGN_FLAG   -> value.toLong() < 0
 		IA32Processor.FlagType.ZERO_FLAG   -> value == ULong.MIN_VALUE
 		IA32Processor.FlagType.PARITY_FLAG -> value.countOneBits() % 2 == 0
 		else                               -> TODO("Unsupported flag: $flag")
